--- ece4/experiment-config-cpld-old.yml
@@ -1,73 +1,0 @@
-- base.context:
-    experiment:
-      id: TEST
-      description: A new ECE4 coupled experiment
-
-      schedule:
-        all: !rrule >
-          DTSTART:19900101
-          RRULE:FREQ=YEARLY;UNTIL=19920101
-
-      nemo:
-        initial_state: null
-        start_from_restart: null
-
-      monitoring:
-        activate: false
-
-      force_from_scratch: false
-
-    model_config:
-      components: [oifs, nemo, rnfm, xios, oasis]
-
-      oifs:
-        grid: !noparse "{{model_config.oifs.all_grids.TL63L31}}"
-
-      nemo:
-        grid: !noparse "{{model_config.nemo.all_grids.ORCA2L31}}"
-
-      oasis:
-        copy_weights: false
-
-    job:
-
-      # Configure launch method with settings
-      # supported methods:
-      #   slurm-hetjob  # SLURM heterogeneous job
-      #   slurm-mp-taskset  # SLURM srun with multi-prog, hostfile and taskset
-      #   slurm-shell  # SLURM + generic shell script template
-      #                # needs job.launch.shell.script: <shell_script>
-      launch:
-        method: slurm-shell
-        shell:
-          script: run-srun-multiprog.sh
-
-      resubmit: true
-
-      oifs:
-        ntasks: 132
-        ntasks_per_node: 44
-        omp_num_threads: 1
-        omp_stacksize: "128M"
-      nemo:
-        ntasks: 132
-        ntasks_per_node: 44
-      xios:
-        ntasks: 1
-        ntasks_per_node: 1
-
-      slurm:
-        sbatch:
-          opts:
-            account: "spitdav2"
-            time: "06:00:00"
-            output: !noparse "{{experiment.id}}.log"
-            job-name: !noparse "ECE4_{{experiment.id}}"
-            qos: "np"
-            ntasks-per-core: 1
-        srun:
-            args: [
-            --label,
-            --kill-on-bad-exit,
-        ]
-
