--- ece4/user-config-example.yml
@@ -1,19 +1,0 @@
-- base.context:
-    experiment:
-      run_dir:  !noparse "RUNDIR/{{experiment.id}}"
-      base_dir: !noparse "BASEDIR"
-      src_dir:  !noparse "{{experiment.base_dir}}/sources"
-      ini_dir:  !noparse "INIDIR"
-
-      monitoring:
-        # To activate monitoring, set
-        #   experiment.monitoring.activated: true
-        # in the experiment configuration.
-        # In order to upload the monitoring results to Redmine, replace
-        # REDMINE_API_KEY below with your real Redmine API key from
-        # https://dev.ec-earth.org/my/account
-        redmine_api_key: REDMINE_API_KEY
-
-        # If redmine_api_key is not a valid key (unset or REDMINE_API_KEY), then
-        # a Markdown representation of monitoring results is created locally in
-        # the monitoring/markdown directory.
